# Backend/cmd/app/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is an online version of the \"The Message\" board game backend API",
        "title": "The Message API",
        "contact": {}
    },
    "host": "127.0.0.1:8080",
    "paths": {
        "/api/v1/games": {
            "post": {
                "description": "Start a new game",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "games"
                ],
                "summary": "Start a new game",
                "parameters": [
                    {
                        "description": "Players",
                        "name": "players",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/request.CreateGameRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/request.CreateGameResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/player_cards/{id}": {
            "get": {
                "description": "GetPlayerCardsByPlayerId",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "player_cards"
                ],
                "summary": "GetPlayerCards",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Player ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/request.PlayerCardsResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "request.CreateGameRequest": {
            "type": "object",
            "properties": {
                "players": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/request.PlayerInfo"
                    }
                }
            }
        },
        "request.CreateGameResponse": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "string"
                },
                "token": {
                    "type": "string"
                }
            }
        },
        "request.PlayerCardsResponse": {
            "type": "object",
            "properties": {
                "cards": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                }
            }
        },
        "request.PlayerInfo": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                }
            }
        }
    }
}